ref: isolate any splitting functions for Slots, Tasks, Steps, and Timeline into a separate module "splitter"
