+++ src/models/step/impls.rs
@@ -1,12 +1,5 @@
-use super::{NewStep, Step, StepStatus};
-use crate::{
-    errors::Error,
-    models::{
-        goal::{Goal, Tag},
-        slot::Slot,
-        timeline::Timeline,
-    },
-};
+use super::{Step, StepStatus};
+use crate::models::{goal::Tag, slot::Slot};
 use std::cmp::Ordering;
 
 impl PartialEq for Step {
@@ -91,48 +84,6 @@ impl Step {
 
     pub fn get_slots(&self) -> Vec<Slot> {
         self.slots.clone()
-    }
-
-    /// Split a Step into list of Steps based on given Step duration.
-    /// - Note: This function will change below in the resulted steps:
-    ///     - Step.status = StepStatus::ReadyToSchedule
-    ///     - Step.tags = empty list
-    pub fn split(&mut self, counter: &mut usize) -> Result<Vec<Step>, Error> {
-        // TODO 2023-06-22: Debug notes: This function not clone step.start and step.deadline
-        if self.duration == 1 {
-            // && !self.tags.contains(&Tag::DoNotSort) {
-            return Err(Error::CannotSplit);
-        }
-        let mut steps = Vec::new();
-        let timeline = Timeline {
-            slots: self.get_slots().into_iter().collect(),
-        };
-        let goal = Goal {
-            id: self.goal_id.clone(),
-            title: self.title.clone(),
-            tags: self.tags.clone(),
-            after_goals: self.after_goals.clone(),
-            ..Default::default()
-        };
-        let new_step = NewStep {
-            step_id: *counter,
-            title: self.title.clone(),
-            duration: 1,
-            goal,
-            timeline,
-            status: StepStatus::Uninitialized,
-            timeframe: None,
-        };
-
-        for _ in 0..self.duration {
-            let mut step = Step::new(new_step.clone());
-            step.id = *counter;
-            step.status = StepStatus::ReadyToSchedule;
-            step.tags = vec![];
-            *counter += 1;
-            steps.push(step);
-        }
-        Ok(steps)
     }
 
     /// Remove conflicted step slots with a given slot [slot_to_remove]
@@ -325,78 +276,6 @@ mod tests {
             let expected_step_slot = Slot::mock(Duration::hours(8), 2023, 01, 03, 3, 0);
 
             assert_eq!(step.slots[0], expected_step_slot);
-        }
-    }
-
-    mod split {
-        use chrono::Duration;
-
-        use crate::models::{
-            slot::Slot,
-            step::{Step, StepStatus},
-        };
-
-        #[test]
-        fn test_split() {
-            let duration: usize = 3;
-            let mut counter: usize = 1;
-
-            let goal_timeframe = Slot::mock(Duration::days(5), 2023, 6, 1, 0, 0);
-            let mut step = Step::mock(
-                "test",
-                duration,
-                0,
-                StepStatus::ReadyToSchedule,
-                vec![goal_timeframe],
-                None,
-            );
-            let steps = step.split(&mut counter).unwrap();
-            dbg!(&step, &steps);
-
-            let mut expected_steps = vec![
-                Step::mock(
-                    "test",
-                    1,
-                    0,
-                    StepStatus::ReadyToSchedule,
-                    vec![goal_timeframe],
-                    None,
-                ),
-                Step::mock(
-                    "test",
-                    1,
-                    0,
-                    StepStatus::ReadyToSchedule,
-                    vec![goal_timeframe],
-                    None,
-                ),
-                Step::mock(
-                    "test",
-                    1,
-                    0,
-                    StepStatus::ReadyToSchedule,
-                    vec![goal_timeframe],
-                    None,
-                ),
-            ];
-            expected_steps[1].id = 2;
-            expected_steps[2].id = 3;
-            dbg!(&expected_steps);
-
-            assert_eq!(steps, expected_steps);
-            assert_eq!(counter, 4);
-
-            assert_eq!(steps[0].id, expected_steps[0].id);
-            assert_eq!(steps[1].id, expected_steps[1].id);
-            assert_eq!(steps[2].id, expected_steps[2].id);
-
-            assert_eq!(steps[0].duration, expected_steps[0].duration);
-            assert_eq!(steps[1].duration, expected_steps[1].duration);
-            assert_eq!(steps[2].duration, expected_steps[2].duration);
-
-            assert_eq!(steps[0].status, expected_steps[0].status);
-            assert_eq!(steps[1].status, expected_steps[1].status);
-            assert_eq!(steps[2].status, expected_steps[2].status);
         }
     }
 }
