+++ src/services/step_generator.rs
@@ -23,38 +23,6 @@ impl Step {
             after_goals: new_step.goal.after_goals,
         }
     }
-
-    /// When a step duration exceeded threshold, it will be splitted
-    /// into 1 hour steps
-    pub fn apply_duration_threshold(&self) -> Vec<Step> {
-        let threshold: usize = 8;
-        let mut new_step = self.clone();
-
-        if new_step.duration > 0 && new_step.duration < threshold {
-            dbg!(&new_step);
-            vec![new_step]
-        } else {
-            // Make first Step with duration as threshold,
-            // then split steps into 1 hours steps till finish
-            // remaining_duration
-
-            let mut steps: Vec<Step> = Vec::new();
-            let mut first_step = new_step.clone();
-            first_step.duration = threshold;
-            dbg!(&first_step);
-            steps.push(first_step);
-            dbg!(&steps);
-
-            let mut remaining_duration = new_step.duration - threshold;
-            new_step.duration = remaining_duration;
-            let new_steps_splitted = new_step.split(&mut remaining_duration).unwrap();
-            dbg!(&new_steps_splitted);
-            steps.extend(new_steps_splitted);
-            dbg!(&steps);
-
-            steps
-        }
-    }
 }
 
 impl Goal {
@@ -134,209 +102,7 @@ impl Goal {
 
 #[cfg(test)]
 mod tests {
-
     mod goal {
-        mod apply_duration_threshold {
-            use chrono::Duration;
-
-            use crate::models::{
-                goal::Goal,
-                slot::Slot,
-                step::{NewStep, Step, StepStatus},
-                timeline::Timeline,
-            };
-
-            #[test]
-            fn test_duration_less_8_hrs() {
-                let duration: usize = 7;
-                let timeframe = Slot::mock(Duration::days(5), 2023, 6, 1, 0, 0);
-
-                let new_step = NewStep {
-                    step_id: 1,
-                    title: "test".to_string(),
-                    duration,
-                    goal: Goal::mock("1", "test", timeframe.clone()),
-                    timeline: Timeline::new(),
-                    status: StepStatus::ReadyToSchedule,
-                    timeframe: Some(timeframe),
-                };
-                let new_step = Step::new(new_step);
-                dbg!(&new_step);
-
-                let generated_steps = new_step.apply_duration_threshold();
-                dbg!(&generated_steps);
-
-                let expected_step = Step::mock(
-                    "test",
-                    7,
-                    0,
-                    StepStatus::ReadyToSchedule,
-                    vec![timeframe],
-                    None,
-                );
-
-                assert_eq!(generated_steps, vec![expected_step.clone()]);
-                assert_eq!(generated_steps[0].id, expected_step.id);
-                assert_eq!(generated_steps[0].duration, expected_step.duration);
-                assert_eq!(generated_steps[0].status, expected_step.status);
-            }
-
-            /// Test Step::apply_duration_threshold when goal.min_duration > 8 hours
-            /// ```markdown
-            /// =========================
-            /// Input:
-            /// Goal {
-            ///    id: "1",
-            ///    title: "test",
-            ///    min_duration: Some(
-            ///        10,
-            ///    ),
-            ///    max_duration: None,
-            ///    budgets: None,
-            ///    repeat: None,
-            ///    start: Some(
-            ///        2023-06-01T00:00:00,
-            ///    ),
-            ///    deadline: Some(
-            ///        2023-06-06T00:00:00,
-            ///    ),
-            ///    tags: [],
-            ///    filters: None,
-            ///    children: None,
-            ///    after_goals: None,
-            ///}
-            ///
-            /// ===========================
-            /// Output:
-            /// expected_step = [
-            ///    Step {
-            ///        id: 1,
-            ///        goal_id: "1",
-            ///        title: "test",
-            ///        duration: 8,
-            ///        status: ReadyToSchedule,
-            ///        flexibility: 0,
-            ///        start: None,
-            ///        deadline: None,
-            ///        slots: [
-            ///            Slot {
-            ///                start:   2023-06-01 00,
-            ///                 end:    2023-06-06 00,
-            ///            },
-            ///        ],
-            ///        tags: [],
-            ///        after_goals: None,
-            ///    },
-            ///    Step {
-            ///        id: 2,
-            ///        goal_id: "1",
-            ///        title: "test",
-            ///        duration: 1,
-            ///        status: ReadyToSchedule,
-            ///        flexibility: 0,
-            ///        start: None,
-            ///        deadline: None,
-            ///        slots: [
-            ///            Slot {
-            ///                start:   2023-06-01 00,
-            ///                 end:    2023-06-06 00,
-            ///            },
-            ///        ],
-            ///        tags: [],
-            ///        after_goals: None,
-            ///    },
-            ///    Step {
-            ///        id: 3,
-            ///        goal_id: "1",
-            ///        title: "test",
-            ///        duration: 1,
-            ///        status: ReadyToSchedule,
-            ///        flexibility: 0,
-            ///        start: None,
-            ///        deadline: None,
-            ///        slots: [
-            ///            Slot {
-            ///                start:   2023-06-01 00,
-            ///                 end:    2023-06-06 00,
-            ///            },
-            ///        ],
-            ///        tags: [],
-            ///        after_goals: None,
-            ///    },
-            ///
-            ///]
-            ///
-            ///
-            /// ```
-            #[test]
-            fn test_duration_more_8_hrs() {
-                let duration: usize = 10;
-                let timeframe = Slot::mock(Duration::days(5), 2023, 6, 1, 0, 0);
-                let timeline = Timeline {
-                    slots: vec![timeframe.clone()].into_iter().collect(),
-                };
-
-                let new_step = NewStep {
-                    step_id: 1,
-                    title: "test".to_string(),
-                    duration,
-                    goal: Goal::mock("1", "test", timeframe.clone()),
-                    timeline,
-                    status: StepStatus::ReadyToSchedule,
-                    timeframe: None,
-                };
-                let new_step = Step::new(new_step);
-                dbg!(&new_step);
-                let generated_steps = new_step.apply_duration_threshold();
-                dbg!(&generated_steps);
-
-                let mut expected_steps = vec![
-                    Step::mock(
-                        "test",
-                        8,
-                        0,
-                        StepStatus::ReadyToSchedule,
-                        vec![timeframe],
-                        None,
-                    ),
-                    Step::mock(
-                        "test",
-                        1,
-                        0,
-                        StepStatus::ReadyToSchedule,
-                        vec![timeframe],
-                        None,
-                    ),
-                    Step::mock(
-                        "test",
-                        1,
-                        0,
-                        StepStatus::ReadyToSchedule,
-                        vec![timeframe],
-                        None,
-                    ),
-                ];
-                expected_steps[1].id = 2;
-                expected_steps[2].id = 3;
-                dbg!(&expected_steps);
-
-                assert_eq!(generated_steps, expected_steps);
-                assert_eq!(generated_steps.len(), 3);
-
-                assert_eq!(generated_steps[0].id, expected_steps[0].id);
-                assert_eq!(generated_steps[1].id, expected_steps[1].id);
-                assert_eq!(generated_steps[2].id, expected_steps[2].id);
-
-                assert_eq!(generated_steps[0].duration, expected_steps[0].duration);
-                assert_eq!(generated_steps[1].duration, expected_steps[1].duration);
-                assert_eq!(generated_steps[2].duration, expected_steps[2].duration);
-
-                assert_eq!(generated_steps[0].status, expected_steps[0].status);
-                assert_eq!(generated_steps[1].status, expected_steps[1].status);
-                assert_eq!(generated_steps[2].status, expected_steps[2].status);
-            }
-        }
-
         mod generate_steps {
             use chrono::Duration;
