+++ src/tests/timeline/mod.rs
@@ -47,33 +47,3 @@ fn test_get_next() {
         assert!(false);
     }
 }
-
-#[test]
-fn test_split_into_days() {
-    let init_year = 2022;
-    let init_month = 1;
-    let init_day = 1;
-    let hour: u32 = 0;
-    let minute: u32 = 0;
-    let days_count: i64 = 5;
-    let duration = Duration::days(days_count);
-
-    let timeline = Timeline::mock(duration, init_year, init_month, init_day);
-
-    let expected_result = Timeline {
-        slots: vec![
-            Slot::mock(Duration::days(1), init_year, init_month, 1, hour, minute),
-            Slot::mock(Duration::days(1), init_year, init_month, 2, hour, minute),
-            Slot::mock(Duration::days(1), init_year, init_month, 3, hour, minute),
-            Slot::mock(Duration::days(1), init_year, init_month, 4, hour, minute),
-            Slot::mock(Duration::days(1), init_year, init_month, 5, hour, minute),
-        ]
-        .into_iter()
-        .collect(),
-    };
-
-    let splitted_timeline = timeline.get_split_into_days();
-
-    dbg!(&splitted_timeline);
-    assert_eq!(expected_result, splitted_timeline);
-}
